refactor: 게임 진행 및 결과 출력 로직 하나의 파일로 병합 - 분리되어있던 출력 파일 하나의 파일로 병합한다.
